Renaming Login to Authentication for clarity, some error handling and formatting
