+++ backend/src/authentication/two-factor.controller.ts
@@ -3,18 +3,18 @@ import { TwoFactorService } from './two-factor.service';
 
 @Controller('two-factor')
 export class TwoFactorController {
-constructor(private readonly twoFactorService: TwoFactorService) {}
+  constructor(private readonly twoFactorService: TwoFactorService) {}
 
   @Get('qrcode')
   async qrcode() {
-	const qrcode = await this.twoFactorService.getQRCode();
-	return qrcode
+    const qrcode = await this.twoFactorService.getQRCode();
+    return qrcode
   }
 
   @Post('verify')
   async verify(@Body() body: { oneTimePassword: string }) {
-	const { oneTimePassword } = body;
-	const verified = await this.twoFactorService.verifyOneTimePassword(oneTimePassword);
-	return verified
+    const { oneTimePassword } = body;
+    const verified = await this.twoFactorService.verifyOneTimePassword(oneTimePassword);
+    return verified
   }
 }
