+++ frontend/src/components/Authentication/ProtectedRoute.tsx
