+++ frontend/src/components/Authentication/TwoFactor.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import axios from 'axios';
 
-
 const Login2FA = () => {
 	const [twoFactorAuthenticationEnabled, setTwoFactorAuthenticationEnabled] = useState(false);
 	const [qrcodeUrl, setQrcodeUrl] = useState(null);
